+++ shuup/xtheme/__init__.py
@@ -66,9 +66,11 @@ class XThemeAppConfig(AppConfig):
         ],
         "admin_module": [
             "shuup.xtheme.admin_module:XthemeAdminModule",
+            "shuup.xtheme.admin_module:XthemeFontsAdminModule",
             "shuup.xtheme.admin_module:XthemeSnippetsAdminModule",
         ],
         "xtheme_resource_injection": ["shuup.xtheme.resources:inject_global_snippet"],
+        "admin_template_injector": ["shuup.xtheme.admin_injector.XthemeAdminTemplateInjector"],
     }
 
     def ready(self):