+++ shuup/xtheme/admin_module/views/__init__.py
@@ -8,6 +8,11 @@
 from shuup.xtheme.admin_module.views._snippet import SnippetDeleteView, SnippetEditView, SnippetListView
 from shuup.xtheme.admin_module.views._theme import (
     ActivationForm,
+    AdminThemeConfigDetailView,
+    AdminThemeForm,
+    FontEditView,
+    FontForm,
+    FontListView,
     TemplateView,
     ThemeConfigDetailView,
     ThemeConfigView,
@@ -17,6 +22,11 @@ from shuup.xtheme.admin_module.views._theme import (
 
 __all__ = [
     "ActivationForm",
+    "FontEditView",
+    "AdminThemeForm",
+    "AdminThemeConfigDetailView",
+    "FontForm",
+    "FontListView",
     "SnippetDeleteView",
     "SnippetEditView",
     "SnippetListView",