+++ shuup/xtheme/forms.py
@@ -8,10 +8,18 @@
 import warnings
 from copy import deepcopy
 from django import forms
+from django.forms.widgets import TextInput
+from django.urls.base import reverse_lazy
 from django.utils.translation import ugettext_lazy as _
 
+from shuup.admin.forms.fields import Select2ModelField
+from shuup.admin.forms.quick_select import QuickAddRelatedObjectSelect
+from shuup.admin.forms.widgets import FileDnDUploaderWidget
+from shuup.admin.shop_provider import get_shop
 from shuup.utils.deprecation import RemovedInFutureShuupWarning
-from shuup.xtheme.models import ThemeSettings
+from shuup.xtheme.models import Font, ThemeSettings
+
+from .models import AdminThemeSettings
 
 
 class GenericThemeForm(forms.ModelForm):
@@ -69,3 +77,72 @@ class GenericThemeForm(forms.ModelForm):
             )
         self.instance.update_settings(self.cleaned_data)
         return self.instance
+
+
+class FontForm(forms.ModelForm):
+    class Meta:
+        model = Font
+        fields = "__all__"
+
+    def __init__(self, *args, **kwargs):
+        self.request = kwargs.pop("request")
+        super(FontForm, self).__init__(*args, **kwargs)
+        self.fields["woff"].widget = FileDnDUploaderWidget(upload_path="/admin_typography/", clearable=True)
+        self.fields["woff2"].widget = FileDnDUploaderWidget(upload_path="/admin_typography/", clearable=True)
+        self.fields["ttf"].widget = FileDnDUploaderWidget(upload_path="/admin_typography/", clearable=True)
+        self.fields["svg"].widget = FileDnDUploaderWidget(upload_path="/admin_typography/", clearable=True)
+        self.fields["eot"].widget = FileDnDUploaderWidget(upload_path="/admin_typography/", clearable=True)
+
+    def save(self, commit=True):
+        self.instance.shop = get_shop(self.request)
+        return super(FontForm, self).save(commit)
+
+
+class QuickAddFontSelect(QuickAddRelatedObjectSelect):
+    url = reverse_lazy("shuup_admin:xtheme.font.new")
+    model = "xtheme.Font"
+
+
+class AdminThemeForm(forms.ModelForm):
+    class Meta:
+        model = AdminThemeSettings
+        fields = "__all__"
+        labels = {
+            "primary_color": _("Choose the primary color:"),
+            "secondary_color": _("Choose the secondary color:"),
+            "text_color": _("Choose the primary text color:"),
+            "success_color": _("Choose the success (green) style primary color:"),
+            "danger_color": _("Choose the danger (red) style primary color:"),
+        }
+        widgets = {
+            "primary_color": TextInput(attrs={"type": "color"}),
+            "secondary_color": TextInput(attrs={"type": "color"}),
+            "text_color": TextInput(attrs={"type": "color"}),
+            "success_color": TextInput(attrs={"type": "color"}),
+            "danger_color": TextInput(attrs={"type": "color"}),
+        }
+
+    def __init__(self, *args, **kwargs):
+        super(AdminThemeForm, self).__init__(*args, **kwargs)
+
+        if self.instance.pk:
+            initial_header_font = self.instance.admin_header_font
+            initial_body_font = self.instance.admin_body_font
+        else:
+            initial_header_font = kwargs.get("initial", {}).get("admin_header_font")
+            initial_body_font = kwargs.get("initial", {}).get("admin_body_font")
+
+        self.fields["admin_header_font"] = Select2ModelField(
+            label=_("Admin Header Font"),
+            initial=initial_header_font,
+            model=Font,
+            required=False,
+            widget=QuickAddFontSelect(editable_model=Font, initial=initial_header_font),
+        )
+        self.fields["admin_body_font"] = Select2ModelField(
+            label=_("Admin Body Font"),
+            initial=initial_body_font,
+            model=Font,
+            required=False,
+            widget=QuickAddFontSelect(editable_model=Font, initial=initial_body_font),
+        )